Reimplements moveup!, movedown!
Amusingly, this turns out not to need invoke, or the menu reference
to the cursor, because the overloads of request mean we know that
the return value of `moveup!` and `movedown!` will be assigned to
that Ref.

QuikTest will need the Ref, however, and it's the right affordance
to offer to other user applications.